requirements.txt
certifi==2018.10.15
numpy==1.15.2
pandas==0.23.4
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
